requirements.txt
streamlit==1.35.0
fastapi==0.95.0
uvicorn==0.22.0
beautifulsoup4==4.12.3
textblob==0.17.1
vaderSentiment==3.3.2
gtts==2.2.3
requests==2.31.0
lxml==5.2.1
python-dotenv==1.0.0
groq

